Update the typehint for the delete_empty options
As per the [Symfony docs], the argument for the delete_empty callable
should be nullable. We got an HTTP 500 when submitting a form. It was
traced back to null being passed to the callable, but the typehint not
allowing it.

[Symfony
docs](https://symfony.com/doc/current/reference/forms/types/collection.html#delete-empty)